refact: abstract poopulateDefaultValues
